+++ NEWS.md
@@ -1,4 +1,4 @@
-# 0.21.3.9000
+# 0.21.4
 
 * If a study has no features table for a modelID, a study can still have
 linkouts to external resources in the results table, as long as the linkouts